requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
numpy
pandas
python-dotenv
faiss-cpu
num2words
matplotlib
plotly
scipy
scikit-learn
